requirements.txt
customtkinter
pillow
requests